requirements/base.txt
# Bleeding edge Django
django==1.9.2

# Configuration
django-environ==0.4.0
django-secure==1.0.1


# Forms
django-braces==1.8.1
django-crispy-forms==1.6.0

# Models
django-model-utils==2.3.1

# Images
Pillow==3.0.0# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.23.0

# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2015.6# Redis support
django-redis==4.2.0
redis>=2.10.0
celery==3.1.18

# Your custom requirements go here
djangorestframework==3.3.2
sodapy==1.2.0
django-phonenumber-field==1.0.0
django-rest-swagger==0.3.5
tweepy==3.5.0
django-facebook==6.0.3
python-instagram==1.3.2
easy-thumbnails==2.3
django-image-cropping==1.0.3
django-filter==0.12.0
django-push-notifications==1.4.1
channels==0.11.0
asgi-redis==0.10.0
